pyproject.toml
[project]
name = "MyCommonLib"
version = "0.8.11"
description = "Common library"
authors = [{name= "Romolo Politi", email= "romolo.politi@inaf.it"}]
license = "GPL-3.0-or-later"
readme = "README.md"
keywords = ["SETUP LIBS"]
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"
]



requires-python = ">=3.10,<4.0"

dependencies=[
"rich (>=13.9.4)",
"pyyaml (>=6.0.2)",
"setuptools (>=80.9.0)",
"semantic-version-tools (>=1.4.0,<2.0.0)",
]

[tool.poetry.urls]
Repository = "https://github.com/RomoloPoliti-INAF/myCommonLib"
"Bug Tracker" = "https://github.com/RomoloPoliti-INAF/myCommonLib/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
include = ["CHANGELOG.md"]
packages=[{include = "MyCommonLib", from="src"}]




[dependency-groups]
devel = [
    "pytest (>=9.0.2,<10.0.0)"
]
